Refactored manifest filter function
