+++ pkg/cliutil/install/chart.go
@@ -13,19 +13,12 @@ import (
 
 const YamlDocumentSeparator = "\n---\n"
 
-// This type represents a function that can be used to filter and transform the contents of a manifest.
-// It returns three values:
-//   - skip: if true, the input manifest will be excluded from the output
-//   - content: if skip is false, this value will be included in the output manifest
-//   - err: if != nil, the whole manifest retrieval operation will fail
-type ManifestFilterFunc func(manifest manifest.Manifest) (skip bool, content string, err error)
-
 // Renders the content of a the Helm chart archive located at the given URI.
 //   - chartArchiveUri: location of the chart, this can be either an http(s) address or a file path
 //   - valueFileName: if provided, the function will look for a value file with the given name in the archive and use it to override chart defaults
 //   - renderOptions: options to be used in the render
-//   - manifestFilter: used to filter and transform the contents of the manifest
-func GetHelmManifest(chartArchiveUri, valueFileName string, opts renderutil.Options, filterFunc ManifestFilterFunc) ([]byte, error) {
+//   - manifestFilter: a collection of functions that can be used to filter and transform the contents of the manifest. Will be applied in the given order.
+func GetHelmManifest(chartArchiveUri, valueFileName string, opts renderutil.Options, filterFunctions ...ManifestFilterFunc) ([]byte, error) {
 
 	// Download chart archive
 	chartFile, err := cliutil.GetResource(chartArchiveUri)
@@ -51,27 +44,20 @@ func GetHelmManifest(chartArchiveUri, valueFileName string, opts renderutil.Opti
 		return nil, err
 	}
 
-	// Collect rendered manifests
-	var validManifests []string
-	for _, manifestFile := range tiller.SortByKind(manifest.SplitManifests(renderedTemplates)) {
-		manifestContent := manifestFile.Content
+	manifests := tiller.SortByKind(manifest.SplitManifests(renderedTemplates))
 
-		// Apply filter function, if provided
-		if filterFunc != nil {
-			skip, content, err := filterFunc(manifestFile)
-			if err != nil {
-				return nil, errors.Wrapf(err, "reading value file")
-			}
-			if skip {
-				continue
-			}
-			manifestContent = content
-		}
-
-		validManifests = append(validManifests, manifestContent)
+	// Apply filter functions to manifests
+	for _, filterFunc := range filterFunctions {
+		manifests, err = filterFunc(manifests)
 	}
 
-	return []byte(strings.Join(validManifests, YamlDocumentSeparator)), nil
+	// Collect manifests
+	var manifestsContent []string
+	for _, m := range manifests {
+		manifestsContent = append(manifestsContent, m.Content)
+	}
+
+	return []byte(strings.Join(manifestsContent, YamlDocumentSeparator)), nil
 }
 
 // Searches for the value file with the given name in the chart and returns its raw content.
